requirements_fla.txt
streamlit
openpyxl
anthropic
python-dotenv
Pillow
PyPDF2
